requirements/aws.txt
aioboto3
aiobotocore
requests-auth-aws-sigv4
s3fs
dill